input: Rework input plugin
* Upgrade to bevy_mod_picking 0.13 (Patched git version)
* Create custom selection component and handling
* Restructure input related things into module
