+++ Cargo.toml
@@ -16,7 +16,7 @@ opt-level = 3
 
 [dependencies]
 bevy_asset_loader = "0.16"
-bevy_mod_picking = { version = "0.12" }
+bevy_mod_picking = { version = "0.13.0", git = "https://github.com/keis/bevy_mod_picking", branch = "ui-fix" }
 bevy_obj = "0.10.1"
 bevy = { version = "0.10", features = ["wayland", "dynamic_linking"] }
 clap = {version = "4.3.0", features = ["derive"] }
