+++ src/lib.rs
@@ -13,7 +13,6 @@ pub mod material;
 pub mod party;
 pub mod path;
 pub mod scene;
-pub mod selection;
 pub mod slide;
 pub mod structure;
 pub mod turn;
